requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
channels==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
cssbeautifier==1.15.1
daphne==4.1.0
Django==5.0.3
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-js-asset==2.2.0
djlint==1.34.1
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jsbeautifier==1.15.1
json5==0.9.25
kombu==5.3.7
mysqlclient==2.2.4
pathspec==0.12.1
pillow==10.3.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
